Programming exercise for problem 1 subproblem C
Added code in Haskell for problem 1. Both algorithms (union of convex
hulls and checking for an intersection) are implemented.

The union of convex hull is Graham-Scan algorithm, where we take
points of both convex hulls, sort them, pick a starting point and run
the algorithm.

The intersection is done with a infinite ray from each point in S1 -
we check how many times each point crosses S2. If the number is even,
it is not inside, if it is odd, it is inside, which means that they
intersect.